requirements.txt
requests
pyspark
pymongo
python-dotenv
pyyaml
dnspython